requirements.txt
# =============================================================================
# DEPRECATION WARNING:
#
# The file `requirements.txt` does not influence the package dependencies and
# will not be automatically created in the next version of PyScaffold (v4.x).
#
# Please have look at the docs for better alternatives
# (`Dependency Management` section).
# =============================================================================
#
# Add your pinned requirements so that they can be easily installed with:
# pip install -r requirements.txt
# Remember to also add them in setup.cfg but unpinned.
# Example:
# numpy==1.13.3
# scipy==1.0
#

google-cloud-storage==1.33.0
matplotlib==3.1.0
numpy==1.18.2
opencv-python==4.2.0.34
pandas==1.0.3
PyQt5==5.13.0
pytest==5.0.1
pytest-cov==2.10.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.8.1
Sphinx==2.4.0
